Undo renaming as lagacyViews
