+++ packages/server/src/api/controllers/view/index.ts
@@ -1,2 +1,2 @@
-export * as v1 from "./legacyViews"
-export * as v2 from "./views"
+export * as v1 from "./views"
+export * as v2 from "./viewsV2"
